Move runtime methods to runtime.go
