+++ proplet/service.go
@@ -13,7 +13,6 @@ import (
 	"sync"
 	"time"
 
-	pkgerrors "github.com/absmach/propeller/pkg/errors"
 	"github.com/absmach/propeller/proplet/api"
 	"github.com/tetratelabs/wazero"
 	wazeroapi "github.com/tetratelabs/wazero/api"
@@ -40,84 +39,6 @@ type ChunkPayload struct {
 	ChunkIdx    int    `json:"chunk_idx"`
 	TotalChunks int    `json:"total_chunks"`
 	Data        []byte `json:"data"`
-}
-
-type WazeroRuntime struct {
-	runtimes map[string]wazero.Runtime
-	modules  map[string]wazeroapi.Module
-	mutex    sync.Mutex
-}
-
-func (w *WazeroRuntime) StartApp(ctx context.Context, appName string, wasmBinary []byte, functionName string) (wazeroapi.Function, error) {
-	if appName == "" {
-		return nil, fmt.Errorf("start app: appName is required but missing: %w", pkgerrors.ErrMissingValue)
-	}
-	if len(wasmBinary) == 0 {
-		return nil, fmt.Errorf("start app: Wasm binary is empty: %w", pkgerrors.ErrInvalidValue)
-	}
-	if functionName == "" {
-		return nil, fmt.Errorf("start app: functionName is required but missing: %w", pkgerrors.ErrMissingValue)
-	}
-
-	w.mutex.Lock()
-	defer w.mutex.Unlock()
-
-	if _, exists := w.modules[appName]; exists {
-		return nil, fmt.Errorf("start app: app '%s' is already running: %w", appName, pkgerrors.ErrAppAlreadyRunning)
-	}
-
-	runtime := wazero.NewRuntime(ctx)
-
-	module, err := runtime.Instantiate(ctx, wasmBinary)
-	if err != nil {
-		runtime.Close(ctx)
-
-		return nil, fmt.Errorf("start app: failed to instantiate Wasm module for app '%s': %w", appName, pkgerrors.ErrModuleInstantiation)
-	}
-
-	function := module.ExportedFunction(functionName)
-	if function == nil {
-		module.Close(ctx)
-		runtime.Close(ctx)
-
-		return nil, fmt.Errorf("start app: function '%s' not found in Wasm module for app '%s': %w", functionName, appName, pkgerrors.ErrFunctionNotFound)
-	}
-
-	w.modules[appName] = module
-	if w.runtimes == nil {
-		w.runtimes = make(map[string]wazero.Runtime)
-	}
-	w.runtimes[appName] = runtime
-
-	return function, nil
-}
-
-func (w *WazeroRuntime) StopApp(ctx context.Context, appName string) error {
-	if appName == "" {
-		return fmt.Errorf("stop app: appName is required but missing: %w", pkgerrors.ErrMissingValue)
-	}
-
-	w.mutex.Lock()
-	defer w.mutex.Unlock()
-
-	module, moduleExists := w.modules[appName]
-	runtime, runtimeExists := w.runtimes[appName]
-
-	if !moduleExists || !runtimeExists {
-		return fmt.Errorf("stop app: app '%s' is not running: %w", appName, pkgerrors.ErrAppNotRunning)
-	}
-
-	if err := module.Close(ctx); err != nil {
-		return fmt.Errorf("stop app: failed to close module for app '%s': %w", appName, err)
-	}
-	if err := runtime.Close(ctx); err != nil {
-		return fmt.Errorf("stop app: failed to close runtime for app '%s': %w", appName, err)
-	}
-
-	delete(w.runtimes, appName)
-	delete(w.modules, appName)
-
-	return nil
 }
 
 func NewService(ctx context.Context, cfg Config, wasmFilePath string, logger *slog.Logger) (*PropletService, error) {
